Added an endpoint to print the students and course report

+++ Coursera_Exercise/DB_Views/CourseDetails.cs
@@ -4,8 +4,8 @@
     {
         public int Course_id { get; set; }
         public string Course_name {  get; set; }
-        public int Total_time {  get; set; }
-        public int Credit {  get; set; }
+        public short Total_time {  get; set; }
+        public short Credit {  get; set; }
         public string Instructor_name { get; set; }
 
     }
